database was changed

+++ order-service/src/main/java/ru/liga/orderservice/controller/RestaurantMenuItemController.java
@@ -6,22 +6,25 @@ import org.springframework.web.bind.annotation.*;
 import ru.liga.dto.RestaurantMenuItemsDTO;
 import ru.liga.orderservice.service.RestaurantMenuItemService;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequiredArgsConstructor
 public class RestaurantMenuItemController {
 
     private final RestaurantMenuItemService restaurantMenuItemService;
+
     @GetMapping("/restaurantItem/{id}")
     public RestaurantMenuItemsDTO getRestaurantItemById(@PathVariable("id") Long id) {
         return restaurantMenuItemService.getMenuItemById(id);
     }
 
-//    @DeleteMapping("/restaurantItem/{id}")
+    //    @DeleteMapping("/restaurantItem/{id}")
 //    public void deleteOrderItemById(@PathVariable("id") Long id) {
 //        restaurantMenuItemService.deleteMenuItemById(id);
 //    }
     @PutMapping("/restaurantItem/{id}")
-    public ResponseEntity<?> changeResItemPrice(@PathVariable("id") Long id, @RequestBody Integer price) {
-        return restaurantMenuItemService.changeItemPrice(id,price);
+    public ResponseEntity<?> changeResItemPrice(@PathVariable("id") Long id, @RequestBody Double price) {
+        return restaurantMenuItemService.changeItemPrice(id, price);
     }
 }
